updated key_store logic and add min_heap implementation
